+++ src/bl64-bsh-setup.bash
@@ -20,13 +20,14 @@ function bl64_bsh_setup() {
   [[ -z "$BL64_VERSION" ]] && echo 'Error: bashlib64-module-core.bash must be sourced at the end' && return 21
 
   # shellcheck disable=SC2034
-  _bl64_lib_module_is_imported 'BL64_DBG_MODULE' &&
+  _bl64_lib_module_is_imported 'BL64_CHECK_MODULE' &&
+    _bl64_lib_module_is_imported 'BL64_DBG_MODULE' &&
     bl64_dbg_lib_show_function &&
-    _bl64_lib_module_is_imported 'BL64_CHECK_MODULE' &&
     _bl64_lib_module_is_imported 'BL64_FMT_MODULE' &&
     _bl64_lib_module_is_imported 'BL64_XSV_MODULE' &&
     _bl64_lib_module_is_imported 'BL64_TXT_MODULE' &&
     _bl64_lib_module_is_imported 'BL64_FS_MODULE' &&
+    _bl64_bsh_set_options &&
     _bl64_bsh_set_version &&
     BL64_BSH_MODULE="$BL64_VAR_ON"
   bl64_check_alert_module_setup 'bsh'
@@ -61,4 +62,25 @@ function _bl64_bsh_set_version() {
   bl64_dbg_lib_show_vars 'BL64_BSH_VERSION_BASH'
 
   return 0
+}
+
+#######################################
+# Create command sets for common options
+#
+# Arguments:
+#   None
+# Outputs:
+#   STDOUT: None
+#   STDERR: None
+# Returns:
+#   0: always ok
+#######################################
+function _bl64_bsh_set_options() {
+  bl64_dbg_lib_show_function
+
+  # Default XDG paths. Paths may not exist
+  BL64_BSH_XDG_PATH_CONFIG="${HOME}/.config"
+  BL64_BSH_XDG_PATH_CACHE="${HOME}/.cache"
+  BL64_BSH_XDG_PATH_LOCAL="${HOME}/.local"
+  BL64_BSH_XDG_PATH_BIN="${BL64_BSH_XDG_PATH_LOCAL}/bin"
 }